Address ESM Compile Issues
This addresses numorous issues with the npm varient of the package. Everything around env variables, assets, css, translations, and even theming. This added new scripts to help improve this process. This is not the ideal way to handle this, but because we are trying to support both the CDN and NPM versions, this is a decent way to achieve it.